hermes: Create CommandApdu and ResponseApdu classes
Future CLs separate the esim functionality between SGP.22-specific and
modem-specific functionality. This allows for the use of hermes with
different modems simply by creating a new modem implementation. The
fragmentation of APDUs, however, is not functionality specific to any
particular modem.

This CL adds classes to represent command and response APDUs. The
CommandApdu class takes care of functionality such as the creation of
the in-memory structure of an APDU, and APDU fragmentation both for
standard and extended-length APDUs. The ResponseApdu class encapsulates
the reading of response status bytes.

BUG=chromium:847619
TEST=- Manually tested with CL:1199598 and ensured that Bound Profile
       Package was successfully received.
     - All unit tests are passing.

Change-Id: I2393528765f10a819bd6c94e4d58f74bbfde230f
Reviewed-on: https://chromium-review.googlesource.com/1368640
Commit-Ready: Alex Khouderchah <akhouderchah@chromium.org>
Tested-by: Alex Khouderchah <akhouderchah@chromium.org>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>