crosh: Add vmc container command, vsh container shell
This adds support for getting a terminal in a container directly from
crosh via 'vsh <vm> <container>'. It also adds a container subcommand to
vmc which allows creating and/or starting a container and then ending up
in a terminal inside the container when done. Optionally an image server
and image alias may be passed as well to specify an alternate container
image to use.

BUG=None
TEST=Manually tested with crosh
CQ-DEPEND=CL:1252541

Change-Id: Ie8f2ad2166c1ebaff79027bf233b10df2c6ba292
Reviewed-on: https://chromium-review.googlesource.com/1252328
Commit-Ready: Jeffrey Kardatzke <jkardatzke@google.com>
Tested-by: Jeffrey Kardatzke <jkardatzke@google.com>
Reviewed-by: Mike Frysinger <vapier@chromium.org>