vm_tools: cicerone: Don't replace mapped ports in URL requests
If we attempt to open a URL in the container that points to localhost or
loopback, then we were replacing that with the container IP or
penguin.linux.test. If we are the default container, and the port for
that URL request is part of our existing mappings, then do not replace
the hostname since localhost or loopback will work in that scenario.

BUG=chromium:890506
TEST=xdg-open http://localhost:[8000|8001|8000/|8001/] work properly

Change-Id: Iea4a04453600ee4e9931ff5a71fbda0bae5eb60c
Reviewed-on: https://chromium-review.googlesource.com/1252331
Commit-Ready: Jeffrey Kardatzke <jkardatzke@google.com>
Tested-by: Jeffrey Kardatzke <jkardatzke@google.com>
Reviewed-by: Stephen Barber <smbarber@chromium.org>