usb: Align the starting position for cropping
For octopus the starting position(upper-left position) for cropping is
(0, 243).
It has 2 streams with resolution 2592x1944 and 1280x720.
We need to crop the image from 2592x1944 to 2592x1458 and scale the
2592x1458 to 1280x720.
The center crop y is (1944-1458)/2 = 243.
Crop at odd Y position is not correct.
It maps UV plane to Y plane with one pixel shift on Y-axis.
Compare to the original image, we interleaves one right stride and one
wrong stride.

BUG=b:115278653
TEST=Verify there is no UV planes mirrors on preview of octopus.

Change-Id: I0a763f78042c234b01b441dc28e717ddd92ed56e
Reviewed-on: https://chromium-review.googlesource.com/1256522
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Tested-by: Hsu Wei-Cheng <mojahsu@chromium.org>
Reviewed-by: Ricky Liang <jcliang@chromium.org>