shill: cellular: add IPv6 whitelist
In the past, IPv6 has been unconditionally disabled on cellular due to
issues with suspend/resume in which a device receives a new IPv6
address on resume but has applications which continue using the old IPv6
address.

This change does not address that issue, but simply whitelists the
L850-GL modem for use with IPv6. The change is unmodified from
@ejcaruso's original CL:740577.

BUG=chromium:726815
TEST=Deploying `USE=dhcpv6 emerge-nocturne shill` leads to the DUT
     getting an IPv6 device on the wwan interface. Putting the DUT to
     sleep and resuming indeed leads to a cycle of connecting and
     disconnecting to the LTE network for a few minutes before
     connecting.

Change-Id: I7d71a42dc53b0f695692e4eab02b36fa34046b80
Reviewed-on: https://chromium-review.googlesource.com/1319981
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Tested-by: Alex Khouderchah <akhouderchah@chromium.org>
Reviewed-by: Alex Khouderchah <akhouderchah@chromium.org>