hammerd: lock RW after update if RO is locked
When RO section is locked, EC won't jump to RW section unless it is
locked as well.  Therefore, when we are in RO section and we don't need
to update RW, we check if we need to lock RW section, if we do, one
additional reboot is expected.

BUG=b:117303686
TEST=manual on device

Change-Id: I4ee26b2ffdbc1745a1de24686b38aef884b71162
Reviewed-on: https://chromium-review.googlesource.com/1263738
Commit-Ready: Wei-Han Chen <stimim@chromium.org>
Tested-by: Wei-Han Chen <stimim@chromium.org>
Reviewed-by: Joel Kitching <kitching@chromium.org>