arc-setup: Do not abort when arcbootcontinue command failed
As mentioned in the code comment, the command fails when it is
executed after the container is shut down (or in the middle of its
shutdown or crash.) The command can fail, for example, if the
Chrome browser process crashes right after sign in, or if Chrome
browser restarts the container under severe OOM conditions.

Since the command failure is not something we can directly fix,
uploading arc-setup crash reports to the crash server is not
something actionable. It only confuses people looking at crash/.

This CL stops sending such reports, and instead just leave logs
in /var/log/arc.log.

BUG=b:80499056
TEST=ARC still works
Change-Id: Icade3e710821a072913289c3ba2a2e083a21c7e3
Reviewed-on: https://chromium-review.googlesource.com/1266610
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Tested-by: Yusuke Sato <yusukes@chromium.org>
Reviewed-by: Hidehiko Abe <hidehiko@chromium.org>