chaps: Persist token_reinitialized flag file to emit a log message on next daemon startup
chapsd does not have permission to write into /var/log, so it can not
write the flag file that was introduced in CL:1255826.
To be still able to know if a token has been reinitialized while syslog
was not available, chapsd will write a flag file into /var/lib/chaps,
and then emit a log message and delete the flag file on the next start.

BUG=chromium:844537
TEST=manual

Change-Id: Ica1e46014821bd15947290c9c36acebd814de97b
Reviewed-on: https://chromium-review.googlesource.com/1267395
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Tested-by: Pavol Marko <pmarko@chromium.org>
Reviewed-by: Andrey Pronin <apronin@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>