vm_tools: 9s: Migrate from crosvm repo
Just like the p9 crate, 9s was only in crosvm because we didn't have a
proper place for it.  Now that we have proper support for building rust
code move this into vm_tools so that we can build it as part of the
vm_host_tools package.

To make the transition easier, remove the dependency on `sys_util` and
instead depend on the `log` crate from upstream.  Also add a log::Log
implementation using libc's syslog functionality, all of which is
contained in syslog.rs.

BUG=none
TEST='FEATURES=test emerge-{eve,kevin} vm_host_tools`

Change-Id: Iae0fc058702925338948e830d69fdc4cb73e1d67
Signed-off-by: Chirantan Ekbote <chirantan@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1279120
Reviewed-by: Dylan Reid <dgreid@chromium.org>