cryptohome: do not assume we have credentials for all PCAs
It is possible that a database was migrated when we didn't have an
endorsement credential anymore and therefore only the existing
encrypted one was migrated. This makes IsAttestationPrepared()
fail as it assumeswe have encrypted credentials for all PCAs.

This fixes this and also will fail enrollment requests creation if
an encrypted credential is missing.

BUG=chromium:894470
TEST=cros_workon_make --board=${BOARD} --test cryptohome

Change-Id: I4eaaaa840da42f0e4b010a3fcd0b017704866fd0
Reviewed-on: https://chromium-review.googlesource.com/1277900
Commit-Ready: Louis Collard <louiscollard@chromium.org>
Tested-by: Yves Arrouye <drcrash@chromium.org>
Reviewed-by: Andrey Pronin <apronin@chromium.org>