vm_tools: sommelier: fix stride0 on create_prime_buffer
This fixes an issue with using virtio-gpu resources over virtio-wayland
in which the guest and the host would have different values for the
stride of a framebuffer. This does not affect rendering but does confuse
the host compositor when it tries to import the real dmabuf with the
wrong stride. To solve the issue, sommelier attempts to fix the stride
being sent to the compositor over the wayland protocol.

TEST=eglgears_wayland
BUG=chromium:875998,chromium:892242

Change-Id: I31e7c221a941ebebc099b5e57f175951be2008b6
Reviewed-on: https://chromium-review.googlesource.com/1279274
Commit-Ready: Zach Reizner <zachr@chromium.org>
Tested-by: Zach Reizner <zachr@chromium.org>
Reviewed-by: David Riley <davidriley@chromium.org>