Stop sending obsolete property to ARC++.
The supervision transition property is now being used to
identify supervised users, and so, the previous property used
for the identification of those users can be removed.

BUG=b:111802968
TEST=Check that ArcBootContinue does not receive the removed property,
which can be observed by looking at the ArcBootContinue logs and
ensuring the "Unknown option" is no longer displayed.

Change-Id: I1b95a3a54fe5f785f22da8aa271e0e46a8612f0d
Reviewed-on: https://chromium-review.googlesource.com/1284869
Commit-Ready: Pedro Simonetti Garcia <pedrosimonetti@chromium.org>
Tested-by: Pedro Simonetti Garcia <pedrosimonetti@chromium.org>
Reviewed-by: Yusuke Sato <yusukes@chromium.org>
Reviewed-by: Bruno Kim Medeiros Cesar <brunokim@chromium.org>
Reviewed-by: Yury Khmel <khmel@chromium.org>