shill: add test for wifi/cell cumulative metrics
The test checks that the update and report callbacks
for cumulative wifi an cellular times are behaving
as expected.  It does *not* check that the callbacks
are automatically invoked as expected: that is tested
in the cumulative metrics unit test.

(Thanks to Kirtika for showing how to deal with mocks and
inheritance and other help.)

BUG=chromium:832145
TEST=test completes correctly

Change-Id: I1e2d5c47640e82f7b5b8790b5e9f1bf085a8450c
Reviewed-on: https://chromium-review.googlesource.com/1286228
Commit-Ready: Luigi Semenzato <semenzato@chromium.org>
Tested-by: Luigi Semenzato <semenzato@chromium.org>
Reviewed-by: Ben Chan <benchan@chromium.org>