vm_tools: Add functionality to update guest time.
For now this can only be invoked manually, with `concierge_client --set_time`.
This call will update the time in all VMs to the current host time.
Invoking this code automatically on resume (and potentially periodically) is
deferred to a future CL.

This RPC is necessary because Linux on ARM does not have paravirtualized
time, so (especially on resume) there can be significant clock drift as
compared to the host.  We work around this on ARM by triggering guest
clock update on resume, using this RPC.

BUG=chromium:823406
TEST=Unit tests (in this CL); manually ran concierge_client call and
verified that guest time was updated.  In a future CL I will add tests
in src/platform/tast-tests/src/chromiumos/tast/local/bundles/cros/vm/subtest/.

Change-Id: Ieb46889ad0bad975c4a360af12e45097f92b5ab8
Reviewed-on: https://chromium-review.googlesource.com/1286087
Commit-Ready: Miriam Zimmerman <mutexlox@chromium.org>
Tested-by: Miriam Zimmerman <mutexlox@chromium.org>
Reviewed-by: Jeffrey Kardatzke <jkardatzke@google.com>
Reviewed-by: Chirantan Ekbote <chirantan@chromium.org>