shill: net: Allow for the destruction of TCP and UDP sockets
This change allows for sockets to be destroyed on the basis of protocol
(either TCP or UDP) and source IP address. The NetlinkSockDiag
implementation is largely unchanged from ejcaruso's original CL:592690,
which was already functional (although it was unexpected that
SOCK_DESTROY's behavior for TCP was to perform the TCP termination
handshake rather than just to immediately destroy the socket).

BUG=chromium:726815
TEST=Manual testing involving the destruction of TCP and UDP sockets
     using our interface's IPv4 and IPv6 addresses. `ss -tu` shows the
     immediate destruction of UDP sockets, and the forced initiation of
     the TCP termination handshake on TCP sockets.

Change-Id: I868cf65aacec5dde85f173008b0b5db14e4bbf47
Reviewed-on: https://chromium-review.googlesource.com/1319982
Commit-Ready: Alex Khouderchah <akhouderchah@chromium.org>
Tested-by: Alex Khouderchah <akhouderchah@chromium.org>
Reviewed-by: Alex Khouderchah <akhouderchah@chromium.org>