init: Correctly clobber-state on entry to developer mode
A narrow logic hole exists on the very first time that chromiumos enters
developer mode, when the TPM is unowned and the install_attributes file
exists.  Reorder the logic such that a wipe is triggered.

BUG=b:112369945, b:117163103, b:113345814
BUG=chromium:701292
TEST=On Grunt.Careena hardware, construct a recovery image with this
change, and install it.  Enter developer mode, and observe that the
stateful partition is clobbered exactly once with a 5 min wait on the
first reboot.  Exit developer mode, and observe that the stateful
partition is clobbered exactly once without an extra wait.

Change-Id: I39df9d311fde9241aa3b33bb0e012826b2d294cd
Signed-off-by: Jonathan Brandmeyer <jbrandmeyer@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1289632
Commit-Ready: Mattias Nissler <mnissler@chromium.org>
Reviewed-by: Simon Glass <sjg@chromium.org>
Reviewed-by: Andrey Pronin <apronin@chromium.org>