tpm_manager: Fix the random fault in SessionFailure unit test
Since the ReadSpace/WriteSpace does not always enable the session
at the beginning of function, it will reply on the publicInfo
which is returned by GetNVSpacePublicArea(). So we have to mock
the GetNVSpacePublicArea() in SessionFailure test or it will fault
randomly in my environment.

BUG=None
TEST=unit tests in tpm_manager

Change-Id: If8af5d66646f33c9b05f246725e4a4fc4df3cb7c
Reviewed-on: https://chromium-review.googlesource.com/1297209
Commit-Ready: Meng-Huan Yu <menghuan@chromium.org>
Tested-by: Meng-Huan Yu <menghuan@chromium.org>
Reviewed-by: Meng-Huan Yu <menghuan@chromium.org>