qmi2cpp: ebnf_cc: Basic framework and EBNF parsing
This commit contains the basic framework for ebnf_cc startup, as well
as a functional EBNF parser implemented as a recursive-descent parser.

Please note that the EBNF parser is currently not very robust. The
provided EBNF file must be written in standard EBNF (an important
consideration, as many standards use a non-standard variant of EBNF),
and typos in the EBNF file will not generate very clear error messages.
The best way of debugging a parse failure of the EBNF file at this point
is to find the last correctly parsed rule, and look for syntax errors in
the following rule.

BUG=chromium:889672
TEST=`python -m ebnf_cc --verbosity=4 ${EBNF_FILE}` to print out the
      parse tree of the provided EBNF file, which in this case was the
      EBNF specification for ASN.1

Change-Id: I44e835b25b6a32530b19e3fea4d27cb953353e33
Reviewed-on: https://chromium-review.googlesource.com/1299496
Commit-Ready: Alex Khouderchah <akhouderchah@chromium.org>
Tested-by: Alex Khouderchah <akhouderchah@chromium.org>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>