installer: fast_dd: Stop whining about little partitions
When dd-ing partitions across, the installer now attempts to find the
largest block size (up to 2MB) that it can copy correctly. If the
block size it finds is too small, it prints out a big ugly warning.

For small partitions, the speed gains are minimal and not worth
persuing. Quit spewing out that big ugly message for partitions of
128MB or less. This number was chosen entirely arbitrarily.

BUG=None
TEST=scp chromeos-install over, run on Cheza

Change-Id: I6bde672d5eb6328556b9a27b2e403732158dbff6
Reviewed-on: https://chromium-review.googlesource.com/1299754
Commit-Ready: Evan Green <evgreen@chromium.org>
Tested-by: Evan Green <evgreen@chromium.org>
Reviewed-by: Douglas Anderson <dianders@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>