system_api: Add proto fields to support dependency message in uninstall
Per UX workup (slide 8), attempting to uninstall a package that has
dependencies should result in a special error message which lists those
dependencies. Prepare for this by adding the appropriate fields to the
cicerone -> host proto. This is needs to be done first so that I can
prepare the host for the new messages before I modify garcon to send
them.

BUG=chromium:822514
TEST=None

Change-Id: Iede114a239c8d42e7e18846fcddb544f2ee9d811
Reviewed-on: https://chromium-review.googlesource.com/1300197
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Tested-by: Ian Barkley-Yeung <iby@chromium.org>
Reviewed-by: Jeffrey Kardatzke <jkardatzke@google.com>