shill: cellular: Prevent outgoing traffic with stale source IPs
As network operators do their part to improve the internet by deploying
anti-spoofing systems, we see a connectivity issue come up as follows:
   * application connects using socket that gets bound to the
      appropriate current source IP address
   * system is provisioned new IP addresses
   * application continues to use socket, which results in the sending
      of packets with a source IP that we no longer have
   * network operator kicks us off the network for spoofing

This issue is common when our MNO uses dynamic IP allocation, as we can
expect to get a different IP address every time we connect to the
network, regardless of when we were last provisioned a set of IP
addresses.

This CL employs a two-part solution to solve this issue. First, it uses
NetlinkSockDiag to kill all of the TCP sockets on our interface prior to
going to sleep. Secondly, it blackholes all outgoing traffic that does
not match our interface's current IP addresses.

BUG=chromium:726815
TEST=Manually attempted to reproduce the bug after applying this
     change. Previous repro steps no longer cause connectivity issues.

Change-Id: Ib908237ec6355d3147655091ad3769100c6d15c7
Reviewed-on: https://chromium-review.googlesource.com/1319985
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Tested-by: Alex Khouderchah <akhouderchah@chromium.org>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>