arc: Add the rw flag to the ptmx mount for master
This is largely the same as a6c219e940ff82daf5c835c6fc5492aa5e691bfc,
but for the master container instead. Below is the original commit's
description:

This mount is not visible to run_oci when it starts building the
mountpoint list, so it does not know what the original flags for it will
be. This change helps run_oci by providing an explicit "rw" for the
/dev/ptmx mount, so that it ends up with the right set of flags.

BUG=b:117343353
TEST=Manually boot master-arc-dev container, logcat works now

Change-Id: I6954cb3acb2ea675bc3d38c8a8795d4e435cc6f7
Reviewed-on: https://chromium-review.googlesource.com/1302593
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Tested-by: Ben Lin <linben@chromium.org>
Reviewed-by: Yusuke Sato <yusukes@chromium.org>
Reviewed-by: Christopher Morin <cmtm@google.com>
Reviewed-by: Mike Frysinger <vapier@chromium.org>