modemfwd: support compressed firmware files
This CL extends modemfwd to support compressed firmware files. A
'compression' field is added to the main and carrier firmware entry in
the firmware manifest to specify whether a firmware file is uncompressed
or compressed with xz. A compressed firmware file will be decompressed
into a temporary file, which will then be flashed to the modem.

BUG=b:117213846
TEST=Run unit tests.

Change-Id: I45afbccb53afa769e321e859d8dd0dde4e204417
Reviewed-on: https://chromium-review.googlesource.com/1309088
Commit-Ready: Ben Chan <benchan@chromium.org>
Tested-by: Ben Chan <benchan@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>