vm_tools: concierge: Refactor Subnet class
Move Subnet out of SubnetPool and make it its own class so that we can
use it without using SubnetPool.  Also add some more tests and fix a bug
where a subnet with a prefix of 0 would return an AvailableCount() of
2^64 - 1.

There should be no functional difference with this change.

BUG=none
TEST=unit tests and manual

Change-Id: Ibc437ccc73b2d455c3924ba90c916a362f3cf550
Signed-off-by: Chirantan Ekbote <chirantan@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1311175
Reviewed-by: Stephen Barber <smbarber@chromium.org>