cryptohome: Removed unused MountTask object types.
As part of the work to remove MountTask from cryptohome, which will
allow glib to be removed as a dependency, this removes all MountTask
types which are not currently used. The next step is to remove the usage
of the rest from service.cc.

BUG=chromium:811551
TEST=FEATURES="test" emerge-${BOARD} cryptohome

Change-Id: Ia4d09bbb954e4067bcd6c023fbc47e1e65e07cb1
Reviewed-on: https://chromium-review.googlesource.com/1311714
Commit-Ready: Greg Kerr <kerrnel@chromium.org>
Tested-by: Greg Kerr <kerrnel@chromium.org>
Reviewed-by: Andrey Pronin <apronin@chromium.org>