installer: Limit badblocks to 100 errors
When a disk firmware crashes, read to all blocks will fail.
It takes over an hour to report all block of a 32GB SSD.
hardware_diagnostics.log also overflow the recovery image stateful,
and dmesg contains only block access errors.

As soon as there are over 100 errors, we know the drive is bad.

BUG=b:117937174
TEST=Check badblocks accept the command.

Change-Id: I29715d51765f42bbafb8e63a9f39fc1b535f79c9
Signed-off-by: Gwendal Grignou <gwendal@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1313378
Reviewed-by: Mike Frysinger <vapier@chromium.org>