metrics: memd: fix scaling of parameters
Parameters (meaning constant quantities, such as the low-mem
margin) were not scaled in the same way as comparable quantities,
as defined by the "group" attribute.  This fixes the problem.
(Example: the margin and the "available" quantity are now
correctly showing the crossing points, which are the points
at which a low-mem notification is sent.)

BUG=none
TEST=tested visually

Change-Id: I8d8c8b505596364713469d450df37ee3515be81e
Reviewed-on: https://chromium-review.googlesource.com/1313769
Commit-Ready: Luigi Semenzato <semenzato@chromium.org>
Tested-by: Luigi Semenzato <semenzato@chromium.org>
Reviewed-by: Sonny Rao <sonnyrao@chromium.org>