diagnostics: Skip binding non-existing directories
Fix the upstart script for the diagnosticsd daemon to not
request minijail to bind non-existing directories.

This refers to directories that the daemon wishes to read but
aren't crucial to them, and whose presence is outside of the
daemon's control. For example, the number of hwmon virtual
devices and their paths under /sys/devices/ vary depending on
the kernel and drivers versions.

Without this fix, the diagnosticsd daemon couldn't even start
when some of such directories are missing.

BUG=chromium:869377,chromium:899766
TEST=check that the daemon starts successfully

Change-Id: Iee1247da1c1f548f903a67400b1f2bbdad181412
Reviewed-on: https://chromium-review.googlesource.com/1315287
Commit-Ready: Maksim Ivanov <emaxx@chromium.org>
Tested-by: Maksim Ivanov <emaxx@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>