diagnostics: make telem report proc/meminfo
Make the telemetry tool actually communicate with diagnosticsd.
Open a socket and communicate via gRPC. The message is hardcoded
to GetProcDataRequest with type=MEMINFO. The response is printed
to the console. Future CLs will refactor telem to have a smaller
main.cc and move the logic to a static library.

BUG=chromium:901126
TEST=(on DUT)telem, output should be the contents of /proc/meminfo.

Change-Id: I2f8f5feb0ac39f2fcedbf77fd3bd3ea68230ecdc
Signed-off-by: Paul Moy <pmoy@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1324311
Reviewed-by: Maksim Ivanov <emaxx@chromium.org>