shill: Add data structure to contain elements with finite lifetimes
There are various scenarios in which we want to maintain a set of
elements that have finite lifetimes, such as expiring routing table
entries or dns mappings. The TimeoutSet class serves this purpose,
taking care of removing elements from the set when they expire, and
optionally informing the client when elements expire.

BUG=chromium:726815
TEST=- Manual testing with integer and IPAddress element types
     - All unit tests are passing

Change-Id: Ia38cada01343c364acf00d5161f8906aa01ce20a
Reviewed-on: https://chromium-review.googlesource.com/1319983
Commit-Ready: Eric Caruso <ejcaruso@chromium.org>
Tested-by: Alex Khouderchah <akhouderchah@chromium.org>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>