power: allow CellularContoller to listen by both proximity and tablet mode
This commit enables CellularController to use a combination of proximity
and tablet mode to determine whether the antenna power needs to be kept
low or can be raised.

The algorithm chosen is - as always - a conservative one where both inputs
need to be in "HIGH-allowed" mode (FAR for proximity, OFF for tablet mode)
in order to allow HIGH power to be selected.

The selection of which inputs to listen to is controlled by powerd prefs.

BUG=b:113303487
TEST=manual on Nautilus LTE + added test cases

Change-Id: I91dc3af1af783693ab20c7b8dd712786075e96fa
Signed-off-by: Enrico Granata <egranata@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1318274
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Reviewed-by: Dan Erat <derat@chromium.org>