vm_tools: add crostini_client, the one client to rule them all
The concept of crostini_client is to have one helper that interacts with
all the services, such as concierge and cicerone. Initially, it is only
a replacement for crosh's vmc command, but it could be expanded to
replace the other command line clients in vm_tools.

TEST=crostini_client vmc start termina
BUG=None

Change-Id: I48293efd8c67e8020c4ae79f9a9360a3a1ed41ff
Reviewed-on: https://chromium-review.googlesource.com/1322590
Commit-Ready: Zach Reizner <zachr@chromium.org>
Tested-by: Zach Reizner <zachr@chromium.org>
Reviewed-by: Dylan Reid <dgreid@chromium.org>