diagnostics: refactor telem/main.cc into libtelem
telem/main.cc has a lot of logic which should be refactored
into a static library, which the telem tool and telemetry
controller (which doesn't exist yet) can both use. Create
the beginnings of libtelem, and move the logic from
telem/main.cc into libtelem.

BUG=chromium:901126
TEST=build and deploy diagnostics on sarien, then run telem.
Telem should print the contents of /proc/meminfo to the console.

Change-Id: I1236c20aca3125a59516e0bf6d0d0dfcc9f1be25
Signed-off-by: Paul Moy <pmoy@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1324313
Reviewed-by: Maksim Ivanov <emaxx@chromium.org>