diagnostics: Sync mojom handshake-related API from Chrome
Pull changes in diagnosticsd.mojom that were made in Chrome's
repository (http://crrev.com/c/1288273) and adopt the diagnosticsd
implementation of Mojo handshake accordingly.

These new interfaces assume that there's a special interface in the
API that allows to establish full-duplex communication between the
diagnosticsd and the browser (DiagnosticsdServiceFactory).
Previously, the idea was that the handshake occurs by calling an
Init() method on the same interface that is used for other requests.

BUG=chromium:869377
TEST=unit tests

Change-Id: Iafcf0ec582b89006936410e615522bbe6731c692
Reviewed-on: https://chromium-review.googlesource.com/c/1327081
Reviewed-by: Maksim Ivanov <emaxx@chromium.org>
Tested-by: Maksim Ivanov <emaxx@chromium.org>
Trybot-Ready: Maksim Ivanov <emaxx@chromium.org>