metrics: Fix incorrect path for thermal logging
Temperature logging was not working because the path for thermal zones
was incorrect, and because the type strings had trailing newlines
which were causing comparisons to fail.

This trims trailing new lines in type, updates the zone path, and
clarifies variable names to help avoid this mistake in the future.

BUG=chromium:850311
TEST=Deploy and check that metrics are logged in chrome://histograms

Change-Id: Id630ad9902247bfbf26eabaa0cd97e8a0834f1f9
Reviewed-on: https://chromium-review.googlesource.com/1330043
Commit-Ready: Fletcher Woodruff <fletcherw@chromium.org>
Tested-by: Fletcher Woodruff <fletcherw@chromium.org>
Reviewed-by: Dan Erat <derat@chromium.org>