vm_tools: Add parser for TZif files to cicerone
In order to set timezones within containers in a portable way, we need
support for extracting POSIX compliant timezone strings from the
timezones located in /usr/share/zoneinfo.

Adds a simple parser for TZif version 2 and version 3 formats that
extracts the embedded timezone string. It does not support version 1
formats because those do not have the embedded string and are not
generated by newer versions of the timezone-data package.

BUG=chromium:829934
TEST=run unit tests

Change-Id: I7e149c070de46827a2e6a2bce9d88cb59932116d
Reviewed-on: https://chromium-review.googlesource.com/1342659
Commit-Ready: Fletcher Woodruff <fletcherw@chromium.org>
Tested-by: Fletcher Woodruff <fletcherw@chromium.org>
Reviewed-by: Chirantan Ekbote <chirantan@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>