biod: Refactor CrosFpDevice to re-use function in bio_crypto_init
This CL refactors some of CrosFpDevice, including moving the class
declarations inside the CrosFpBiometricsManager header, so that we can
re-use the WaitOnEcBoot() routine, instead of copying it into
bio_crypto_init.

The next step is to break out CrosFpDevice into it's own file and
header, but this CL can be a less-disruptive first step.

BUG=b:117909326, chromium:906164
TEST=Build and boot, verify that enrollment and unlock works as
     expected.

Change-Id: I9ec3042be6f350822e01feebd40d9c57ae8a8417
Reviewed-on: https://chromium-review.googlesource.com/1338210
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Tested-by: Prashant Malani <pmalani@google.com>
Reviewed-by: Nicolas Norvez <norvez@chromium.org>