shill: add HTTPS probe and logging for HTTPS probe
The next step in improving captive portal detection is to add an
HTTPS probe. This change adds the probe but does not use the results
to make any portal diagnoses. We do this to ensure that the probe
does not cause any regressions, and to collect some metrics to track
improvements.

BUG=chromium:903645
TEST=shill unittests

Change-Id: I304c9815624384544adc2b07bfb69f1d738853f6
Reviewed-on: https://chromium-review.googlesource.com/1341226
Commit-Ready: Matthew Wang <matthewmwang@chromium.org>
Tested-by: Matthew Wang <matthewmwang@chromium.org>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Reviewed-by: Ben Chan <benchan@chromium.org>