cros-disks: Recover from empty drivefs datadirs with incorrect owners.
If cros-disks stops after the drivefs datadir is created, but before
it's chowned, future attempts to mount with that datadir are doomed.
Since this should only occur before running anything that should write
to it, attempt to recover from this by deleting the empty datadir and
fall through to creating a new datadir.

BUG=chromium:905946
TEST=unit tests, manual

Change-Id: Ief859d98c0dc61faf3cdb62fce0aa3bbebc19b95
Reviewed-on: https://chromium-review.googlesource.com/1338787
Commit-Ready: Sam McNally <sammc@chromium.org>
Tested-by: Sam McNally <sammc@chromium.org>
Reviewed-by: Anand Mistry <amistry@chromium.org>