authpolicy: Remove auth retry
Removes retries of AuthenticateUser requests in case of network errors.
Those caused a 10 second delay when when a user tried to log in offline.

Auth retries solved a problem when a user logged in while the network
was reconnecting. In this case, the Kerberos ticket cannot be fetched
without retries. After a few seconds in the session, when the network is
available, GetUserStatus() succeeds and show a popup asking the user to
relog since there's no Kerberos ticket.

However, in the meantime we can restore Kerberos tickets from backup
(see chromium:738433), so that the popup won't show up, at least not
for ~10 hours.

In the long run, we'll have to address the issue of reconnecting
networks during login. As a workaround, a user can enforce an online
logon by entering their password wrong a couple of times.

BUG=chromium:856556
TEST=cros_workon_make --board=amd64-generic --test authpolicy
     Went through repro steps of chromium:846725 and made sure it didn't
     regress.

Change-Id: Ib7579f48e1053bd9f10589b72f528ef6d7155e74
Reviewed-on: https://chromium-review.googlesource.com/1340619
Commit-Ready: Lutz Justen <ljusten@chromium.org>
Tested-by: Lutz Justen <ljusten@chromium.org>
Reviewed-by: Roman Sorokin <rsorokin@chromium.org>