cros-disks: whitelist 'nosymfollow' mount option
Allow Chrome to mount filesystems with the 'nosymfollow' option if the
FsNosymfollow Chrome flag is enabled. This flag is default-disabled in
Chrome, but we want the feature to be enable-able for all filesystem
types except for drivefs for now.

BUG=chromium:867807
TEST=unit tests, manual testing similar to autotest in CL:1269737 (plug
in usb and try to traverse symlink on the filesystem in
/media/removable/...)

Change-Id: I7a518dcea3e12796fbd42c2d124f716566693fe9
Reviewed-on: https://chromium-review.googlesource.com/1345312
Commit-Ready: Micah Morton <mortonm@chromium.org>
Tested-by: Micah Morton <mortonm@chromium.org>
Reviewed-by: Ben Chan <benchan@chromium.org>