vm_tools: p9: Add fuzzing support
Add support for fuzzing by exporting a `fuzzing` module that is only
built when the `fuzzing` build config is set.  The fuzzing module
provides public access to internal functions for the purpose of fuzzing
these interfaces.

Also add a new `p9-fuzz` sub-crate that we will use to hold all the fuzz
targets.  Currently there is only one target, which fuzzes the wire
format decoding implementation.

BUG=none
TEST=Run the fuzzer under `cros_fuzz --board eve shell`

Change-Id: Ie4bd05a89746bc4ec25d7de310de976bf87c9b57
Signed-off-by: Chirantan Ekbote <chirantan@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1344830