diagnostics: Use "bytes" for binary proto fields
This is a pure cleanup which replaces the use of "string" with
"bytes" for binary proto fields, as the latter is semantically more
suitable and provides better compatibility on some languages.

No implications for the existing code are expected, since protobuf
C++ bindings just use std::string for both types.

BUG=chromium:869377
TEST=the diagnostics package builds

Change-Id: I0a0f523f11f094c8684739ab2b2da089eac15d54
Reviewed-on: https://chromium-review.googlesource.com/1348649
Commit-Ready: Maksim Ivanov <emaxx@chromium.org>
Tested-by: Maksim Ivanov <emaxx@chromium.org>
Reviewed-by: Oleh Lamzin <lamzin@google.com>
Reviewed-by: Pavol Marko <pmarko@chromium.org>