intel: ipu3: Release last RAW buffer early
Currently to make sure the two pipes outputs are the same,
the last request raw buffer is not released until the current
request processes completely. If the request processing time
is too long, the raw buffers may be consumed quickly so that
the other requests will be droped.

This patch releases the last raw buffer immediately when it
is not used any more.

BUG=b:117583021
TEST=Make sure camera works on soraka.

Change-Id: Icad6b5ec20270d0f2d91c4d0f935293471ed0d05
Signed-off-by: Hongju Wang <hongju.wang@intel.com>
Reviewed-on: https://chromium-review.googlesource.com/1351082
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Tested-by: Wei Meng <wei.meng@intel.com>
Reviewed-by: Ricky Liang <jcliang@chromium.org>
Reviewed-by: Wei Meng <wei.meng@intel.com>