diagnostics: Refactor std::function -> base::Callback conv
Add a helper that builds base::Callback from the given std::function.
Use this helper to replace the code that achieved the same via a
combination of base::Bind and a lambda.

Historically, the older version of libchrome allowed to simply call
base::Bind for that purposes.

BUG=chromium:907528
TEST=the diagnostics package builds

Change-Id: I99ba6882f485726faa85c476d1ff41789823c709
Reviewed-on: https://chromium-review.googlesource.com/1349569
Commit-Ready: Maksim Ivanov <emaxx@chromium.org>
Tested-by: Maksim Ivanov <emaxx@chromium.org>
Reviewed-by: Pavol Marko <pmarko@chromium.org>